requirements.txt
streamlit
mysql-connector-python
bcrypt
pandas
python-telegram-bot==20.7
plyer
